+++ CHANGELOG.md
@@ -14,3 +14,8 @@
 * Updated docs
 * Updated example app icon
 * If no FW match, lookup nearest FW version to use to get the flash file size.
+
+## 0.0.1-dev.15
+
+* Converted to package.
+* Insert not found revision to DB